bot-ai/requirements.txt
aiogram==2.25.1
aiosqlite==0.19.0
python-dotenv==1.0.0
httpx==0.24.1
aiohttp==3.8.5
